Got driver loading done!
Drivers now load. I've changed the format coming out of modexec_export()
to be more in line with the instinctive revulsion to prototypes in Perl.
Also updated all of the error handling, removed the need to use Exporter
in modexec_export(), and switched to Module::Load for loading modules
instead of the string evals I was doing. String evals still happen in
in order to prevent the need to export the modexec_export(). I still
am not in love with that, I will revisit it later.